Bring in the patches from FreeBSD ports.

+++ bsd/swapinternal.h
@@ -10,8 +10,8 @@ int
 BSDInitSwapInfo();
 
 void
-BSDGetSwapInfo(int* total, int* free);
+BSDGetSwapInfo(int64_t* total, int64_t* free);
 #endif
 
 void
-BSDGetSwapCtlInfo(int* total, int* free);
+BSDGetSwapCtlInfo(int64_t* total, int64_t* free);
